renaming of assim jobs
