+++ src/tasks/assim/arome_BS3D__obstype.py
