+++ src/tasks/assim/arome_BSM3DVar_CNT0.py
@@ -5,9 +5,8 @@ from vortex import toolbox
 from vortex.layout.nodes import Driver, Family, LoopFamily
 
 from .raw2odb.batodb import BatorODB
-from .screenings.screening_LAM3D import Screening
+from .screenings.screeningCNT0_LAM3D import Screening as ScreeningCNT0
 from .minims.minimCNT0_LAM3D import Minim as MinimCNT0
-from .minims.minimOOPS_LAM3D import Minim as MinimOOPS
 
 
 def setup(t, **kw):
@@ -19,15 +18,9 @@ def setup(t, **kw):
                         loopconf='rundates',
                         loopsuffix='.{}',
                         nodes=[
-                        Family('BSM', ticket=t, on_error='delayed_fail', nodes=[
                             BatorODB(tag='batodb', ticket=t, **kw),
-                            Family('multitasks', ticket=t, on_error='delayed_fail', nodes=[
-                                Screening(tag='screening', ticket=t, **kw),
-                                # delayed_fail to let the minimOOPS run before raising error
-                                MinimCNT0(tag='minimCNT0', ticket=t, on_error='delayed_fail', **kw),
-                                #MinimOOPS(tag='minimOOPS', ticket=t, **kw),
-                                ], **kw),
-                            ], **kw),
+                            ScreeningCNT0(tag='screeningCNT0', ticket=t, **kw),
+                            MinimCNT0(tag='minimCNT0', ticket=t, **kw),
                         ], **kw),
                     ], **kw),
                 ], **kw),
