+++ src/tasks/conf/belenos.ini
@@ -459,7 +459,7 @@ openmp              = 1
 # ASSIM
 # =====
 
-[OOPSBSM_4D_arpege]
+[arpege_BSM4DVar_OOPS]
 geometry            = geometry(globalupd149)
 test_family         = op_obs_file
 timestep            = 1800
@@ -470,7 +470,7 @@ ntasks              = 16
 nnodes              = 1
 openmp              = 8
 
-[OOPSBA1_4D_arpege]
+[arpege_BA4DVar_OOPS]
 geometry            = geometry(globalupd149)
 timestep            = 1800
 rundate             = 2022012106
@@ -480,7 +480,7 @@ ntasks              = 16
 nnodes              = 1
 openmp              = 8
 
-[CNT0BSM_4D_arpege]
+[arpege_BSM4DVar_CNT0]
 geometry            = geometry(globalupd149)
 timestep            = 1800
 rundate             = 2022012106
@@ -490,7 +490,7 @@ ntasks              = 16
 nnodes              = 1
 openmp              = 8
 
-[BS_4D_arpege__obstype]
+[arpege_BS4D__obstype]
 geometry            = geometry(globalupd149)
 timestep            = 1800
 rundate             = 2022012106
@@ -502,7 +502,7 @@ ntasks              = 16
 nnodes              = 1
 openmp              = 8
 
-[OOPSBSM_3D_arpege]
+[arpege_BSM3DVar_OOPS]
 geometry            = geometry(globalupd149)
 timestep            = 1800
 rundate             = 2022012106
@@ -512,7 +512,7 @@ ntasks              = 16
 nnodes              = 1
 openmp              = 8
 
-[CNT0BSM_3D_arpege]
+[arpege_BSM3DVar_CNT0]
 geometry            = geometry(globalupd149)
 timestep            = 1800
 rundate             = 2022012106
@@ -522,7 +522,7 @@ ntasks              = 16
 nnodes              = 1
 openmp              = 8
 
-[BS_3D_arpege__obstype]
+[arpege_BS3D__obstype]
 geometry            = geometry(globalupd149)
 timestep            = 1800
 # loop on obstypes
@@ -534,7 +534,7 @@ ntasks              = 16
 nnodes              = 1
 openmp              = 8
 
-[BSM_3D_arome]
+[arome_BSM3DVar_CNT0]
 geometry            = geometry(france10km)
 timestep            = 60
 # several rundates to cover the whole span of obstypes
@@ -546,7 +546,7 @@ ntasks              = 16
 nnodes              = 1
 openmp              = 8
 
-[BA1_3D_arome]
+[arome_BA3DVar_OOPS]
 geometry            = geometry(france10km)
 timestep            = 60
 # several rundates to cover the whole span of obstypes
@@ -558,7 +558,7 @@ ntasks              = 16
 nnodes              = 1
 openmp              = 8
 
-[BS_3D_arome__obstype]
+[arome_BS3D__obstype]
 geometry            = geometry(france10km)
 timestep            = 60
 # loop on obstypes
@@ -571,7 +571,7 @@ ntasks              = 16
 nnodes              = 1
 openmp              = 8
 
-[BsC_arpege]
+[arpege_BC]
 geometry            = geometry(globalupd149)
 rundate             = 2017121006
 # profile (ntasks = ntasks/node)
